+++ apps/web/src/app/(with-side-bar)/topic/[id]/layout.tsx
@@ -1,0 +1,17 @@
+interface TopicStickerDetailLayoutProps {
+  children: React.ReactNode;
+  modal: React.ReactNode;
+}
+
+export default function TopicStickerDetailLayout({
+  children,
+  modal,
+}: TopicStickerDetailLayoutProps) {
+  return (
+    <>
+      {children}
+      {modal}
+      <div id="modal-root" />
+    </>
+  );
+}
