+++ apps/web/src/components/(with-side-bar)/topic/add-sticker-dialog.tsx
@@ -40,7 +40,7 @@ function AddStickerDialogContent({ topicId }: AddStickerDialogContentProps) {
 
     await createCustomSticker(body, {
       onSuccess: () => {
-        invalidateQueries([TOPIC.GET_TOPIC_BY_ID, topicId]);
+        invalidateQueries([TOPIC.GET_TOPIC_BOARD_BY_ID, topicId]);
         close();
       },
       onError: (error: Error) => {
