+++ apps/web/src/components/(with-side-bar)/topic/flow-canvas.tsx
@@ -13,10 +13,9 @@ import {
   useReactFlow,
 } from "@xyflow/react";
 
-import { AlertTriangle, Lightbulb, Loader2, Plus } from "lucide-react";
+import { AlertTriangle, Loader2 } from "lucide-react";
 
-import CustomNode from "./custom-node";
-import AddTopicDialog from "../(with-side-bar)/layout/add-topic-dialog";
+import Sticker from "./sticker";
 
 interface FlowCanvasProps {
   isLoading: boolean;
@@ -64,7 +63,7 @@ const FlowCanvas = ({
   const nodeTypes: NodeTypes = useMemo(
     () => ({
       custom: (props: NodeProps) => (
-        <CustomNode
+        <Sticker
           {...props}
           key={`topic-nodes-${id}`}
           topicId={id}
@@ -123,7 +122,7 @@ const FlowCanvas = ({
             </>
           )}
         </div>
-      ) : id ? (
+      ) : (
         <ReactFlow
           nodes={nodes}
           edges={edges}
@@ -137,17 +136,6 @@ const FlowCanvas = ({
         >
           <Background />
         </ReactFlow>
-      ) : (
-        <div className="bg-background absolute inset-0 z-10 flex items-center justify-center">
-          <div className="text-center">
-            <Lightbulb size={64} className="text-muted-foreground mx-auto mb-6 opacity-50" />
-            <h3 className="mb-4 text-xl font-semibold">선택된 토픽이 없습니다</h3>
-            <p className="text-muted-foreground mb-6">토픽을 선택하거나 새 토픽을 생성해보세요</p>
-            <AddTopicDialog>
-              <Plus size={16} />새 토픽 생성
-            </AddTopicDialog>
-          </div>
-        </div>
       )}
     </div>
   );
