+++ apps/web/src/components/(with-side-bar)/topic/sticker/content-sticker.tsx
