+++ apps/web/src/components/(with-side-bar)/topic/sticker/remove-user-sticker-dialog.tsx
@@ -1,37 +1,28 @@
-import { useRouter } from "next/navigation";
-
 import { TOPIC } from "@/constants/topic";
 import { invalidateQueries } from "@/lib/tanstack";
-import { useRemoveTopic } from "@/lib/tanstack/mutation/topic";
-import { revalidatePath } from "@/utils/revalidate";
-import {
-  Button,
-  DialogClose,
-  DialogContent,
-  errorToast,
-  successToast,
-  useDialog,
-} from "@linkyboard/components";
+import { useRemoveCustomSticker } from "@/lib/tanstack/mutation/custom-sticker";
+import { Button, DialogClose, DialogContent, errorToast, useDialog } from "@linkyboard/components";
 
 import { Loader2 } from "lucide-react";
 
-export default function RemoveTopicDialog({ topicId }: { topicId: number }) {
-  const router = useRouter();
-
-  const { mutateAsync, isPending } = useRemoveTopic();
+export default function RemoveUserStickerDialog({
+  customStickerId,
+  topicId,
+}: {
+  customStickerId: number;
+  topicId: string;
+}) {
+  const { mutateAsync, isPending } = useRemoveCustomSticker();
   const { close } = useDialog();
 
   const onRemoveTopic = async () => {
-    await mutateAsync(topicId, {
+    await mutateAsync(customStickerId.toString(), {
       onSuccess: () => {
-        successToast("토픽이 삭제되었어요.");
-        invalidateQueries([TOPIC.GET_ALL_TOPICS]);
-        revalidatePath(`/topic?id=${topicId}`);
-        router.back();
+        invalidateQueries([TOPIC.GET_TOPIC_BOARD_BY_ID, topicId]);
         close();
       },
       onError: () => {
-        errorToast("토픽 삭제에 실패했어요.");
+        errorToast("스티커 삭제에 실패했어요.");
       },
     });
   };
@@ -41,7 +32,7 @@ export default function RemoveTopicDialog({ topicId }: { topicId: number }) {
       <div className="space-y-4">
         <div>
           <h2 className="text-foreground text-lg font-semibold">삭제</h2>
-          <p className="text-muted-foreground text-sm">토픽을 삭제 하시겠습니까?</p>
+          <p className="text-muted-foreground text-sm">스티커를 삭제 하시겠습니까?</p>
         </div>
         <div className="flex justify-end gap-2">
           <Button variant="outline" asChild disabled={isPending}>
