+++ apps/web/src/constants/custom-sticker.ts
@@ -1,3 +1,4 @@
 export const CUSTOM_STICKER = {
   GET_AI_MODELS: "get-ai-models",
+  GET_CUSTOM_STICKER_BY_ID: "get-custom-sticker-by-id",
 };
