+++ apps/web/src/constants/topic.ts
@@ -1,4 +1,5 @@
 export const TOPIC = {
+  GET_TOPIC_BOARD_BY_ID: "get-topic-board-by-id",
   GET_TOPIC_BY_ID: "get-topic-by-id",
   GET_ALL_TOPICS: "get-all-topics",
   GET_ALL_CONTENTS: "get-all-contents",
